refactor: Extract DelegateDocumentBuilderFactory & subtypes
Into separate files
